Removed overlay implementation, keeping persistent notification as entry point for non compose root.
