--- lens-ui/src/androidMain/res/values/ids.xml
@@ -1,4 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <item name="lens_overlay_compose_view" type="id" />
-</resources>
