+++ lens-ui/src/commonMain/kotlin/io/github/farhazulmullick/lenslogger/ui/LensMaterialTheme.kt
@@ -1,0 +1,15 @@
+package io.github.farhazulmullick.lenslogger.ui
+
+import androidx.compose.foundation.isSystemInDarkTheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.darkColorScheme
+import androidx.compose.material3.lightColorScheme
+import androidx.compose.runtime.Composable
+
+@Composable
+internal fun LensMaterialTheme(content: @Composable () -> Unit) {
+    MaterialTheme(
+        colorScheme = if (isSystemInDarkTheme()) darkColorScheme() else lightColorScheme(),
+        content = content,
+    )
+}
